# docs/openapi.yaml
openapi: 3.0.3
info:
  title: KuruOps API
  version: "0.1.0"
  description: >
    REST API for KuruOps, a single-tenant SOC alert/incident-management
    platform (Go backend, PostgreSQL, React frontend). This document covers
    every route mounted by internal/httpserver/router.go: the unauthenticated
    /auth/** login flows, /healthz and /metrics, and the JWT-protected
    /api/v1/** surface (alerts, incidents, playbooks, tags, uploads,
    dashboard, live event stream, and admin-only /settings/** configuration).
    See the repository root README for product context, deployment, and the
    default-admin-first-login flow; see TROUBLESHOOTING.md for operational
    issues. This spec favors completeness of shape (paths, params, request/
    response bodies, status codes) over exhaustive field-level validation
    constraints.
servers:
  - url: http://localhost:8080
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      summary: Local username/password login
      description: >
        Unauthenticated. Rate limited per-IP (router-level) and per-email
        (handler-level). Resolves the deployment's single tenant automatically.
        If the account has TOTP enrolled, a correct password returns
        MfaRequiredResponse instead of LoginResponse -- see /auth/mfa/verify
        for the second step.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        "200":
          description: Login succeeded, or the account requires a second factor
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/LoginResponse'
                  - $ref: '#/components/schemas/MfaRequiredResponse'
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          description: >-
            Invalid credentials. Unlike the shared Unauthorized response
            (issued by the auth middleware via http.Error, plain text),
            this comes from the handler itself via writeError -- the same
            application/json {"error": "..."} shape as every other
            handler-level error in this API.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        "429":
          description: Too many login attempts for this account or IP -- same handler-level JSON error shape as 401 above
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        "500":
          $ref: '#/components/responses/InternalError'
  /auth/mfa/verify:
    post:
      tags: [Auth]
      summary: Second step of a login for a TOTP-enrolled account
      description: >
        Unauthenticated. Consumes pendingToken (single-use) from a prior
        /auth/login response and checks code against the account's enrolled
        TOTP secret -- same opaque-failure discipline as /auth/login itself:
        an unknown/expired/already-used pendingToken and a wrong code are
        both a plain 401, never distinguished. Rate limited per pendingToken,
        same as /auth/password-reset/confirm is per reset token.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                pendingToken: { type: string }
                code: { type: string, description: The 6-digit TOTP code. }
              required: [pendingToken, code]
      responses:
        "200":
          description: Login completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          description: Invalid or expired pending token, or wrong code
          content:
            text/plain:
              schema:
                type: string
        "429":
          description: Too many attempts for this pending token
          content:
            text/plain:
              schema:
                type: string
        "500":
          $ref: '#/components/responses/InternalError'
  /auth/login/ldap:
    post:
      tags: [Auth]
      summary: LDAP login
      description: Unauthenticated. Same rate limiting as local login; validates against the tenant's configured LDAP directory.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        "200":
          description: Login succeeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          description: Invalid credentials
          content:
            text/plain:
              schema:
                type: string
        "429":
          description: Too many login attempts
          content:
            text/plain:
              schema:
                type: string
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Exchange a refresh token for a new access token
      description: Unauthenticated (the refresh token itself is the credential). Rotates the refresh token.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
              required: [refreshToken]
      responses:
        "200":
          description: New token pair issued
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  refreshToken:
                    type: string
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          description: Invalid or expired refresh token
          content:
            text/plain:
              schema:
                type: string
        "500":
          $ref: '#/components/responses/InternalError'
  /auth/saml/metadata:
    get:
      tags: [Auth]
      summary: SAML SP metadata document
      description: Unauthenticated. Serves this tenant's SP metadata XML for IdP configuration.
      security: []
      responses:
        "200":
          description: SP metadata XML
          content:
            application/xml:
              schema:
                type: string
        "500":
          $ref: '#/components/responses/InternalError'
  /auth/saml/login:
    get:
      tags: [Auth]
      summary: Initiate SAML login (redirect to IdP)
      description: Unauthenticated. Redirects the browser to the configured IdP's SSO endpoint.
      security: []
      responses:
        "302":
          description: Redirect to the IdP
        "500":
          $ref: '#/components/responses/InternalError'
  /auth/saml/acs:
    post:
      tags: [Auth]
      summary: SAML Assertion Consumer Service callback
      description: Unauthenticated. The IdP posts the SAML assertion here to complete login.
      security: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                SAMLResponse:
                  type: string
      responses:
        "200":
          description: Login completed (response shape mirrors LoginResponse, rendered/redirected by the SAML handler)
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          description: Invalid SAML assertion
          content:
            text/plain:
              schema:
                type: string
        "500":
          $ref: '#/components/responses/InternalError'
  /auth/password-reset/request:
    post:
      tags: [Auth]
      summary: Request a password reset email
      description: >
        Unauthenticated. Always returns 204 regardless of whether the email
        matches a real, local, active account -- the response never leaks
        account existence.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
              required: [email]
      responses:
        "204":
          description: Request accepted (no content, regardless of outcome)
        "400":
          $ref: '#/components/responses/BadRequest'
        "429":
          description: Too many attempts for this account
          content:
            text/plain:
              schema:
                type: string
        "500":
          $ref: '#/components/responses/InternalError'
  /auth/password-reset/confirm:
    post:
      tags: [Auth]
      summary: Confirm a password reset with a token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
                newPassword:
                  type: string
              required: [token, newPassword]
      responses:
        "204":
          description: Password reset
        "400":
          $ref: '#/components/responses/BadRequest'
        "429":
          description: Too many attempts
          content:
            text/plain:
              schema:
                type: string
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Required on every /api/v1/** route. /auth/** is intentionally
        unauthenticated (that's how a session is obtained in the first
        place), and /healthz and /metrics are unauthenticated for
        infrastructure health checks and scraping. A caller whose
        mustChangePassword flag is still set can reach nothing under
        /api/v1/** except POST /api/v1/account/change-password (see
        RequirePasswordChanged), receiving 403 on everything else until the
        password is changed.
...
  responses:
    BadRequest:
      description: Malformed request body, path, or query parameter
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    InternalError:
      description: Unexpected server-side failure
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      description: Standard handler-level error shape, written by handlers.writeError.
      properties:
        error:
          type: string
      required: [error]
...
    LoginRequest:
      type: object
      properties:
        email: { type: string }
        password: { type: string }
      required: [email, password]
    LoginResponse:
      type: object
      properties:
        token: { type: string }
        refreshToken: { type: string }
        user:
          type: object
          properties:
            id: { type: string }
            email: { type: string }
            name: { type: string }
            role: { type: string }
            mustChangePassword: { type: boolean }
    MfaRequiredResponse:
      type: object
      description: >-
        What POST /auth/login returns instead of LoginResponse when the
        account has TOTP enrolled -- the password was correct, but no
        session exists yet. Present pendingToken alongside a 6-digit code to
        POST /auth/mfa/verify (which returns a normal LoginResponse) to
        complete the login.
      properties:
        mfaRequired: { type: boolean, enum: [true] }
        pendingToken: { type: string }
      required: [mfaRequired, pendingToken]
...